fixing telegram model

+++ internal/telegram/message.go
@@ -9,7 +9,7 @@ type Message struct {
 }
 
 type SendMessageRequest struct {
-	ChatID      int                   `json:"chat_id"`
+	ChatID      int64                 `json:"chat_id"`
 	Text        string                `json:"text"`
 	ReplyMarkup *InlineKeyboardMarkup `json:"reply_markup,omitempty"`
 }
